renames common header
