+++ test/include/q_common/init.h
