+++ test/src/common/cbuf/cbuf.c
@@ -28,7 +28,7 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 
 #include "q_shared.h"
 #include "quake.h"
-#include "common.h"
+#include "q_common.h"
 
 #define MAX_CMD_LEN 0x800
 #define MAX_SZBUF_LEN (4 * MAX_CMD_LEN)
