+++ test/src/common/com/com.c
@@ -30,7 +30,7 @@ along with this program; if not, see <http://www.gnu.org/licenses/>.
 
 #include "q_shared.h"
 #include "quake.h"
-#include "common.h"
+#include "q_common.h"
 
 #define STDC17 GNU_STD17
