+++ test/src/common/cvar/cvar.c
@@ -30,7 +30,7 @@ along with this program; if not, see <http://www.gnu.org/licenses/>.
 
 #include "q_shared.h"
 #include "quake.h"
-#include "common.h"
+#include "q_common.h"
 
 static cvar_t *cvars = NULL;
 static qboolean userinfo_modified = False;
