+++ test/src/common/fs/fs.c
@@ -29,7 +29,7 @@ along with this program; if not, see <http://www.gnu.org/licenses/>.
 
 #include "q_shared.h"
 #include "quake.h"
-#include "common.h"
+#include "q_common.h"
 
 typedef struct {
 	char name[MAX_QPATH];
