+++ test/src/common/quake/quake.c
@@ -27,7 +27,7 @@ along with this program; if not, see <http://www.gnu.org/licenses/>.
 
 #include "q_shared.h"
 #include "quake.h"
-#include "common.h"
+#include "q_common.h"
 
 int Quake_Free (void)
 {
